update: unit testing breaks on error update: add return types

+++ src/statistics/norminv.jl
@@ -13,7 +13,7 @@ Convert probability to a normal distribution with a peak at 0.5.
 
 - `n::Float64`
 """
-function norminv(x::Real)
+function norminv(x::Real)::Float64
 
     n = quantile(Distributions.Normal(), x)
